artrate has interests
It wants each color to be a specific proportion. The proportions were
randomly generated and don't all add up to 1.0. You can't please
everybody!

+++ village/artist/perception_test.go
@@ -38,7 +38,7 @@ func TestPerceiveColor(t *testing.T) {
 
 		got := perceiveColor(im, tt.clr)
 
-		if math.Abs(float64(got)-float64(tt.cnt)/100.0) > 0.001 {
+		if math.Abs(got-float64(tt.cnt)/100.0) > 0.001 {
 			t.Errorf("perceiveColor(im, %v) => %f, but expected %f", tt.clr, got, float64(tt.cnt)/100.0)
 		}
 	}
